Move sync code to the blocking module
Move sync code that interacts with the I2C to a separate module,
in preparation for adding an async module.
