+++ src/blocking/channels.rs
@@ -1,0 +1,97 @@
+use crate::{hal, Channel, Error, Pca9685, Register, channels::{get_register_on, get_register_off}};
+
+impl<I2C, E> Pca9685<I2C>
+where
+    I2C: hal::blocking::i2c::Write<Error = E> + hal::blocking::i2c::WriteRead<Error = E>,
+{
+    /// Set the `ON` counter for the selected channel.
+    ///
+    /// Note that the full off setting takes precedence over the `on` settings.
+    /// See section 7.3.3 "LED output and PWM control" of the datasheet for
+    /// further details.
+    pub fn set_channel_on(&mut self, channel: Channel, value: u16) -> Result<(), Error<E>> {
+        if value > 4095 {
+            return Err(Error::InvalidInputData);
+        }
+        let reg = get_register_on(channel);
+        self.write_double_register(reg, value)
+    }
+
+    /// Set the `OFF` counter for the selected channel.
+    pub fn set_channel_off(&mut self, channel: Channel, value: u16) -> Result<(), Error<E>> {
+        if value > 4095 {
+            return Err(Error::InvalidInputData);
+        }
+        let reg = get_register_off(channel);
+        self.write_double_register(reg, value)
+    }
+
+    /// Set the `ON` and `OFF` counters for the selected channel.
+    ///
+    /// Note that the full off setting takes precedence over the `on` settings.
+    /// See section 7.3.3 "LED output and PWM control" of the datasheet for
+    /// further details.
+    pub fn set_channel_on_off(
+        &mut self,
+        channel: Channel,
+        on: u16,
+        off: u16,
+    ) -> Result<(), Error<E>> {
+        if on > 4095 || off > 4095 {
+            return Err(Error::InvalidInputData);
+        }
+        let reg = get_register_on(channel);
+        self.write_two_double_registers(reg, on, off)
+    }
+
+    /// Set the channel always on.
+    ///
+    /// The turning on is delayed by the value argument.
+    /// Note that the full off setting takes precedence over the `on` settings.
+    ///
+    /// See section 7.3.3 "LED output and PWM control" of the datasheet for
+    /// further details.
+    pub fn set_channel_full_on(&mut self, channel: Channel, value: u16) -> Result<(), Error<E>> {
+        if value > 4095 {
+            return Err(Error::InvalidInputData);
+        }
+        let reg = get_register_on(channel);
+        let value = value | 0b0001_0000_0000_0000;
+        self.write_double_register(reg, value)
+    }
+
+    /// Set the channel always off.
+    ///
+    /// This takes precedence over the `on` settings and can be cleared by setting
+    /// the `off` counter with [`set_channel_off`](struct.Pca9685.html#method.set_channel_off).
+    ///
+    /// See section 7.3.3 "LED output and PWM control" of the datasheet for
+    /// further details.
+    pub fn set_channel_full_off(&mut self, channel: Channel) -> Result<(), Error<E>> {
+        let reg = get_register_off(channel);
+        let value = 0b0001_0000_0000_0000;
+        self.write_double_register(reg, value)
+    }
+
+    /// Set the `ON` and `OFF` counter for each channel at once.
+    ///
+    /// The index of the value in the arrays corresponds to the channel: 0-15.
+    /// Note that the full off setting takes precedence over the `on` settings.
+    /// See section 7.3.3 "LED output and PWM control" of the datasheet for
+    /// further details.
+    pub fn set_all_on_off(&mut self, on: &[u16; 16], off: &[u16; 16]) -> Result<(), Error<E>> {
+        let mut data = [0; 65];
+        data[0] = Register::C0_ON_L;
+        for (i, (on, off)) in on.iter().zip(off).enumerate() {
+            if *on > 4095 || *off > 4095 {
+                return Err(Error::InvalidInputData);
+            }
+            data[i * 4 + 1] = *on as u8;
+            data[i * 4 + 2] = (*on >> 8) as u8;
+            data[i * 4 + 3] = *off as u8;
+            data[i * 4 + 4] = (*off >> 8) as u8;
+        }
+        self.enable_auto_increment()?;
+        self.i2c.write(self.address, &data).map_err(Error::I2C)
+    }
+}
